+++ src/api/hooks/useReports/useReports.jsx
@@ -1,0 +1,18 @@
+import {
+  useCreateReport 
+} from './request';
+
+export const useAuth = () => {
+   
+  const {
+    createReport, isLoading: isCreateReportLoading 
+  } = useCreateReport();
+  
+  return {
+    createReport: {
+      createReport,
+      isCreateReportLoading,
+    },
+  };
+};
+
