+++ src/forms/CreateReport/index.js
@@ -1,0 +1,4 @@
+export {
+  CreateReportForm 
+} from './CreateReportForm';
+
